\name{exampleDataLarge}
\docType{data}
\alias{exampleDataLarge}
\title{summarizedData object on 2800 450k Human Methylation samples}
\description{Summarized 450K dataon 2800 samples. The data can 
be use as a background data set in the interactive application.}
\usage{exampleDataLarge}
\value{Pre-summarizedData object on 2800 450k Human Methylation
  samples.}
\format{summarizedData-object}
\keyword{datasets}
\examples{
data(exampleDataLarge)
data(exampleData)
\dontrun{visualize(exampleData, background=exampleDataLarge)} 
}
  
